+++ sdk/resourcemanager/appcontainers/armappcontainers/models.go
@@ -1161,6 +1161,27 @@ type CookieExpiration struct {
 	TimeToExpiration *string `json:"timeToExpiration,omitempty"`
 }
 
+// CorsPolicy - Cross-Origin-Resource-Sharing policy
+type CorsPolicy struct {
+	// REQUIRED; allowed origins
+	AllowedOrigins []*string `json:"allowedOrigins,omitempty"`
+
+	// allow credential or not
+	AllowCredentials *bool `json:"allowCredentials,omitempty"`
+
+	// allowed HTTP headers
+	AllowedHeaders []*string `json:"allowedHeaders,omitempty"`
+
+	// allowed HTTP methods
+	AllowedMethods []*string `json:"allowedMethods,omitempty"`
+
+	// expose HTTP headers
+	ExposeHeaders []*string `json:"exposeHeaders,omitempty"`
+
+	// max time client can cache the result
+	MaxAge *int32 `json:"maxAge,omitempty"`
+}
+
 // CustomDomain - Custom Domain of a Container App
 type CustomDomain struct {
 	// REQUIRED; Resource Id of the Certificate to be bound to this hostname. Must exist in the Managed Environment.
@@ -1176,7 +1197,7 @@ type CustomDomain struct {
 // CustomDomainConfiguration - Configuration properties for apps environment custom domain
 type CustomDomainConfiguration struct {
 	// Certificate password
-	CertificatePassword []byte `json:"certificatePassword,omitempty"`
+	CertificatePassword *string `json:"certificatePassword,omitempty"`
 
 	// PFX or PEM blob
 	CertificateValue []byte `json:"certificateValue,omitempty"`
@@ -1889,6 +1910,14 @@ type Ingress struct {
 	// Bool indicating if HTTP connections to is allowed. If set to false HTTP connections are automatically redirected to HTTPS
 	// connections
 	AllowInsecure *bool `json:"allowInsecure,omitempty"`
+
+	// Client certificate mode for mTLS authentication. Ignore indicates server drops client certificate on forwarding. Accept
+	// indicates server forwards client certificate but does not require a client
+	// certificate. Require indicates server requires a client certificate.
+	ClientCertificateMode *IngressClientCertificateMode `json:"clientCertificateMode,omitempty"`
+
+	// CORS policy for container app
+	CorsPolicy *CorsPolicy `json:"corsPolicy,omitempty"`
 
 	// custom domain bindings for Container Apps' hostnames.
 	CustomDomains []*CustomDomain `json:"customDomains,omitempty"`